requirements.txt
mysql-connector-python==8.0.28
pillow==9.4.0
python-dotenv==1.0.0
reportlab==3.6.12
PyPDF2==3.0.1
